feat(rename-r3-stage-b4): replication.py -> persistence.py + shim (#87)
R3 Stage B.4 per the plan in ~/.claude/plans/ethereal-forging-cookie.md
and tracking issue #87.

Files:
  - helix_context/replication.py -> helix_context/persistence.py (rename)
  - helix_context/replication.py  (new shim)

Class identifier (ReplicationManager) is unchanged — Stage C may
rename the method-level surface (replicate -> persist).

Verification: 3-path identity + tests/test_server.py + test_genome.py
114/114 pass in 2:06.
